Experiment to turn off congestion control in a datachannel.
Partial reliability SCTP relies on sending a FWD ack point pakcet
once the connection detects that some packets have been lost.
dropped. This is a lot of back and forth, and the FWD ack point and
sacks can get lost/delayed.

In some cases (games with low latency requirements) it only the latest
received packet that matters. SACKING and FWD ack pointing all these
packets seems not very useful.

This is an experiment where I turn off the SCTP congestion control, by
always sending the FWD acck point of TSN X-1 together with the data with
TSN X.

This means the receiver will not care of all previous packets, even if
they arrive or not.

I am not yet 100% sure how to deal with SACKs, as of right now I
always send the highest sack in its own packet, though I might evaluate
adding that minimal SACK to the data packet as well.