Detach buffering and sending
Buffering happened in its own pool detached from the user thread
sending, which problems when the buffer was full, since this would
prevent the connection from recovering by refusing messages and draining
itself, as well as not notify the user thread

Now the used thread requesting a send will throw an exception if
buffering fails.

If the messages has been buffered, another thread will pick it up and
perform the send.